Made an self-introduction document for homework 1.
R Markdown cheat sheet from tools was super helpful. It was cool to learn the more # you add for header, the smaller the font gets. **Text** is bold, *text* is italics. Making unordered bullet lists is intuitive. When inserting a link to an image or URL, you need both [caption](link). For image, add ! before the square brackets.