I added my biography to the README.md file.
First, self introduction part includes my name, program of study at the University of Chicago, nationality, age, and information on my sibling. My name is written in bold and program of study is written in italic.

Second, three of my research interest are introduced. It is presented with format of a list. Most unfamiliar yet important term, donju, is written in italic too.

Third, two of my hobbies are presented. I film youtube to share with my family and love to have a short walk with my puppy. My youtube page link is hyperlinked under the channel name 'Lemon's Diary.' Also, the picture of Geni is attached at the very end.