Fixed a few syntax errors and grammar errors.
To me, following the steps in the guide to get a grip at the GitHub workflow is not too hard so far. Markdown is also easy to understand with the help of the guide website. However, I am still not sure about the difference between Markdown and R Markdown and whether should we do anything to distinguish these two when coding with R.