Homework #1 - Autobiography
I thoroughly went through the Setup tab to set up Git within RStudio, as well as relying on the Homework Workflow page, with a very helpful video demonstration to make sure I got it right.
I was surprised how much I struggled to set up Git and GitHub, as well as feeding it through RStudio. Embarrassingly, I got stuck at the part of the Setup where the shell asked for my password and I was confused as to why I could not see what I was typing haha! I consulted a friend who helped me with this. In addtion, I think I struggled partly because I was stumped with the new vocabularly I had to quickly pick up as I could not attend any of the live sessions of class. I figured out how to add in those textual elements was by searching up Markdown Basics.